DEV-3 Moved files
