+++ src/Shared/excelAdapter/dto/StructureDTO.go
